requirements.txt

langchain>=0.1.0
langchain-huggingface>=0.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
huggingface-hub>=0.20.0
requests>=2.31.0